+++ splitio/api/telemetry.py
@@ -2,7 +2,8 @@
 import logging
 
 from splitio.api.client import HttpClientException
-from splitio.api.commons import headers_from_metadata, record_telemetry, APIException
+from splitio.api.commons import record_telemetry
+from splitio.api import headers_from_metadata, APIException
 from splitio.util.time import get_current_epoch_time_ms
 from splitio.models.telemetry import HTTPExceptionsAndLatencies
