+++ splitio/sync/segment.py
@@ -3,7 +3,8 @@ import time
 import json
 import os
 
-from splitio.api.commons import APIException, FetchOptions
+from splitio.api.commons import FetchOptions
+from splitio.api import APIException
 from splitio.tasks.util import workerpool
 from splitio.models import segments
 from splitio.util.backoff import Backoff
