+++ tests/api/test_httpclient.py
@@ -10,14 +10,15 @@ class HttpClientTests(object):
         response_mock = mocker.Mock()
         response_mock.status_code = 200
         response_mock.text = 'ok'
+        response_mock.headers = {}
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.requests.get', new=get_mock)
         httpclient = client.HttpClient()
-        response = httpclient.get('sdk', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.get('sdk', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.SDK_URL + '/test1',
-            headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
+            client.SDK_URL + 'test1',
+            headers={'Content-Type': 'application/json', 'Authorization': 'Bearer some_api_key', 'h1': 'abc'},
             params={'param1': 123},
             timeout=None
         )
@@ -26,9 +27,9 @@ class HttpClientTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = httpclient.get('events', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.get('events', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.EVENTS_URL + '/test1',
+            client.EVENTS_URL + 'test1',
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
             timeout=None
@@ -42,6 +43,7 @@ class HttpClientTests(object):
         response_mock = mocker.Mock()
         response_mock.status_code = 200
         response_mock.text = 'ok'
+        response_mock.headers = {}
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.requests.get', new=get_mock)
@@ -75,13 +77,14 @@ class HttpClientTests(object):
         response_mock = mocker.Mock()
         response_mock.status_code = 200
         response_mock.text = 'ok'
+        response_mock.headers = {}
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.requests.post', new=get_mock)
         httpclient = client.HttpClient()
-        response = httpclient.post('sdk', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.post('sdk', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.SDK_URL + '/test1',
+            client.SDK_URL + 'test1',
             json={'p1': 'a'},
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
@@ -92,9 +95,9 @@ class HttpClientTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = httpclient.post('events', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.post('events', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.EVENTS_URL + '/test1',
+            client.EVENTS_URL + 'test1',
             json={'p1': 'a'},
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
@@ -109,11 +112,12 @@ class HttpClientTests(object):
         response_mock = mocker.Mock()
         response_mock.status_code = 200
         response_mock.text = 'ok'
+        response_mock.headers = {}
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.requests.post', new=get_mock)
         httpclient = client.HttpClient(sdk_url='https://sdk.com', events_url='https://events.com')
-        response = httpclient.post('sdk', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.post('sdk', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://sdk.com' + '/test1',
             json={'p1': 'a'},
@@ -126,7 +130,7 @@ class HttpClientTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = httpclient.post('events', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = httpclient.post('events', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://events.com' + '/test1',
             json={'p1': 'a'},
@@ -139,9 +143,10 @@ class HttpClientTests(object):
         assert get_mock.mock_calls == [call]
 
 class MockResponse:
-    def __init__(self, text, status):
+    def __init__(self, text, status, headers):
         self._text = text
         self.status = status
+        self.headers = headers
 
     async def text(self):
         return self._text
@@ -158,16 +163,16 @@ class HttpClientAsyncTests(object):
     @pytest.mark.asyncio
     async def test_get(self, mocker):
         """Test HTTP GET verb requests."""
-        response_mock = MockResponse('ok', 200)
+        response_mock = MockResponse('ok', 200, {})
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.aiohttp.ClientSession.get', new=get_mock)
         httpclient = client.HttpClientAsync()
-        response = await httpclient.get('sdk', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.get('sdk', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         assert response.status_code == 200
         assert response.body == 'ok'
         call = mocker.call(
-            client.SDK_URL + '/test1',
+            client.SDK_URL + 'test1',
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
             timeout=None
@@ -175,9 +180,9 @@ class HttpClientAsyncTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = await httpclient.get('events', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.get('events', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.EVENTS_URL + '/test1',
+            client.EVENTS_URL + 'test1',
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
             timeout=None
@@ -189,12 +194,12 @@ class HttpClientAsyncTests(object):
     @pytest.mark.asyncio
     async def test_get_custom_urls(self, mocker):
         """Test HTTP GET verb requests."""
-        response_mock = MockResponse('ok', 200)
+        response_mock = MockResponse('ok', 200, {})
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.aiohttp.ClientSession.get', new=get_mock)
         httpclient = client.HttpClientAsync(sdk_url='https://sdk.com', events_url='https://events.com')
-        response = await httpclient.get('sdk', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.get('sdk', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://sdk.com/test1',
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
@@ -206,7 +211,7 @@ class HttpClientAsyncTests(object):
         assert response.body == 'ok'
         get_mock.reset_mock()
 
-        response = await httpclient.get('events', '/test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.get('events', 'test1', 'some_api_key', {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://events.com/test1',
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
@@ -220,14 +225,14 @@ class HttpClientAsyncTests(object):
 
     async def test_post(self, mocker):
         """Test HTTP POST verb requests."""
-        response_mock = MockResponse('ok', 200)
+        response_mock = MockResponse('ok', 200, {})
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.aiohttp.ClientSession.post', new=get_mock)
         httpclient = client.HttpClientAsync()
-        response = await httpclient.post('sdk', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.post('sdk', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.SDK_URL + '/test1',
+            client.SDK_URL + 'test1',
             json={'p1': 'a'},
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
@@ -238,9 +243,9 @@ class HttpClientAsyncTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = await httpclient.post('events', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.post('events', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
-            client.EVENTS_URL + '/test1',
+            client.EVENTS_URL + 'test1',
             json={'p1': 'a'},
             headers={'Authorization': 'Bearer some_api_key', 'h1': 'abc', 'Content-Type': 'application/json'},
             params={'param1': 123},
@@ -252,12 +257,12 @@ class HttpClientAsyncTests(object):
 
     async def test_post_custom_urls(self, mocker):
         """Test HTTP GET verb requests."""
-        response_mock = MockResponse('ok', 200)
+        response_mock = MockResponse('ok', 200, {})
         get_mock = mocker.Mock()
         get_mock.return_value = response_mock
         mocker.patch('splitio.api.client.aiohttp.ClientSession.post', new=get_mock)
         httpclient = client.HttpClientAsync(sdk_url='https://sdk.com', events_url='https://events.com')
-        response = await httpclient.post('sdk', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.post('sdk', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://sdk.com' + '/test1',
             json={'p1': 'a'},
@@ -270,7 +275,7 @@ class HttpClientAsyncTests(object):
         assert get_mock.mock_calls == [call]
         get_mock.reset_mock()
 
-        response = await httpclient.post('events', '/test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
+        response = await httpclient.post('events', 'test1', 'some_api_key', {'p1': 'a'}, {'param1': 123}, {'h1': 'abc'})
         call = mocker.call(
             'https://events.com' + '/test1',
             json={'p1': 'a'},
