+++ tests/api/test_segments_api.py
@@ -4,7 +4,7 @@ import pytest
 import unittest.mock as mock
 
 from splitio.api import segments, client
-from splitio.api.commons import APIException
+from splitio.api import APIException
 from splitio.api.commons import FetchOptions
 from splitio.client.util import SdkMetadata
 from splitio.engine.telemetry import TelemetryStorageProducer
@@ -16,7 +16,7 @@ class SegmentAPITests(object):
     def test_fetch_segment_changes(self, mocker):
         """Test segment changes fetching API call."""
         httpclient = mocker.Mock(spec=client.HttpClient)
-        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}')
+        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}', {})
         segment_api = segments.SegmentsAPI(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
 
         response = segment_api.fetch_segment('some_segment', 123, FetchOptions())
@@ -65,7 +65,7 @@ class SegmentAPITests(object):
     @mock.patch('splitio.engine.telemetry.TelemetryRuntimeProducer.record_sync_latency')
     def test_segment_telemetry(self, mocker):
         httpclient = mocker.Mock(spec=client.HttpClient)
-        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}')
+        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}', {})
         telemetry_storage = InMemoryTelemetryStorage()
         telemetry_producer = TelemetryStorageProducer(telemetry_storage)
         telemetry_runtime_producer = telemetry_producer.get_telemetry_runtime_producer()
