+++ tests/api/test_splits_api.py
@@ -4,7 +4,7 @@ import pytest
 import unittest.mock as mock
 
 from splitio.api import splits, client
-from splitio.api.commons import APIException
+from splitio.api import APIException
 from splitio.api.commons import FetchOptions
 from splitio.client.util import SdkMetadata
 from splitio.engine.telemetry import TelemetryStorageProducer
@@ -17,7 +17,7 @@ class SplitAPITests(object):
     def test_fetch_split_changes(self, mocker):
         """Test split changes fetching API call."""
         httpclient = mocker.Mock(spec=client.HttpClient)
-        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}')
+        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}', {})
         split_api = splits.SplitsAPI(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
 
         response = split_api.fetch_splits(123, FetchOptions())
@@ -66,7 +66,7 @@ class SplitAPITests(object):
     @mock.patch('splitio.engine.telemetry.TelemetryRuntimeProducer.record_sync_latency')
     def test_split_telemetry(self, mocker):
         httpclient = mocker.Mock(spec=client.HttpClient)
-        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}')
+        httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}', {})
         telemetry_storage = InMemoryTelemetryStorage()
         telemetry_producer = TelemetryStorageProducer(telemetry_storage)
         telemetry_runtime_producer = telemetry_producer.get_telemetry_runtime_producer()
