+++ tests/push/test_segment_worker.py
@@ -3,7 +3,7 @@ import time
 import queue
 import pytest
 
-from splitio.api.commons import APIException
+from splitio.api import APIException
 from splitio.push.segmentworker import SegmentWorker
 from splitio.models.notification import SegmentChangeNotification
