+++ tests/sync/test_events_synchronizer.py
@@ -5,7 +5,7 @@ import time
 import pytest
 
 from splitio.api.client import HttpResponse
-from splitio.api.commons import APIException
+from splitio.api import APIException
 from splitio.storage import EventStorage
 from splitio.models.events import Event
 from splitio.sync.event import EventSynchronizer
